computer_vision_app/pyproject.toml
[project]
name = "computer-vision-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "joblib>=1.5.3",
    "mediapipe>=0.10.33",
    "opencv-python>=4.13.0.92",
    "python-fasthtml>=0.13.1",
    "scikit-learn>=1.8.0",
    "streamlit>=1.55.0",
    "streamlit-webrtc>=0.64.5",
]
